requirements.txt
# Google APIs
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# Gemini LLM
google-generativeai>=0.3.0

# Environment variables
python-dotenv>=1.0.0

# Utilities
httplib2>=0.22.0
